chore(clerk-js): Revamp route mock builders
