+++ packages/clerk-js/src/ui/utils/test/createFixture.tsx
@@ -10,59 +10,29 @@ import { AppearanceProvider } from '../../customizables';
 import { FlowMetadataProvider } from '../../elements';
 import { RouteContext } from '../../router';
 import { InternalThemeProvider } from '../../styledSystem';
-import { createClerkMockContexts } from './createClerkMockContexts';
 import { createClientFixtureHelpers, createEnvironmentFixtureHelpers } from './fixtureHelpers';
 import { createBaseClientJSON, createBaseEnvironmentJSON } from './fixtures';
-import { getInitialFixtureConfig } from './mockConfigs';
+import { mockClerkMethods, mockRouteContextValue } from './mockHelpers';
 
 type UnpackContext<T> = NonNullable<T extends React.Context<infer U> ? U : T>;
 
-const createInitialStateConfigParam = (config: any, baseEnvironment: EnvironmentJSON, baseClient: ClientJSON) => {
+const createInitialStateConfigParam = (baseEnvironment: EnvironmentJSON, baseClient: ClientJSON) => {
   return {
     ...createEnvironmentFixtureHelpers(baseEnvironment),
     ...createClientFixtureHelpers(baseClient),
-    mockRouteNavigate: () => {
-      const mockNavigate = jest.fn();
-      config.routeContext.navigate = mockNavigate;
-      return mockNavigate;
-    },
   };
 };
 
 type FParam = ReturnType<typeof createInitialStateConfigParam>;
 type ConfigFn = (f: FParam) => void;
 
-type DeepJestMocked<T> = T extends object
-  ? {
-      [k in keyof T]: T[k] extends object ? jest.Mocked<T[k]> : T[k];
-    }
-  : T;
-
-const mockClerkMethods = (clerk: LoadedClerk): DeepJestMocked<LoadedClerk> => {
-  const mockProp = <T,>(obj: T, k: keyof T) => {
-    if (typeof obj[k] === 'function') {
-      // @ts-ignore
-      obj[k] = jest.fn();
-    }
-  };
-
-  const mockMethodsOf = (obj: any) => {
-    Object.keys(obj).forEach(k => mockProp(obj, k));
-  };
-  mockMethodsOf(clerk.client.signIn);
-  mockMethodsOf(clerk.client.signUp);
-  mockProp(clerk, 'navigate');
-  return clerk as any as DeepJestMocked<LoadedClerk>;
-};
-
 export const createFixture = (componentName: UnpackContext<typeof ComponentContext>['componentName']) => {
   return async (configFn?: ConfigFn) => {
-    const config = getInitialFixtureConfig();
     const baseEnvironment = createBaseEnvironmentJSON();
     const baseClient = createBaseClientJSON();
 
     if (configFn) {
-      configFn(createInitialStateConfigParam(config, baseEnvironment, baseClient));
+      configFn(createInitialStateConfigParam(baseEnvironment, baseClient));
     }
 
     const environmentMock = new Environment(baseEnvironment);
@@ -77,12 +47,14 @@ export const createFixture = (componentName: UnpackContext<typeof ComponentConte
     const tempClerk = new ClerkCtor(frontendApi);
     await tempClerk.load();
     const clerkMock = mockClerkMethods(tempClerk as LoadedClerk);
-    const { mockedRouteContext } = createClerkMockContexts(config);
+
+    const routerMock = mockRouteContextValue();
 
     const fixtures = {
       clerk: clerkMock,
       signIn: clerkMock.client.signIn,
       signUp: clerkMock.client.signUp,
+      router: routerMock,
     };
 
     const MockClerkProvider = (props: any) => {
@@ -90,7 +62,7 @@ export const createFixture = (componentName: UnpackContext<typeof ComponentConte
       return (
         <CoreClerkContextWrapper clerk={clerkMock}>
           <EnvironmentProvider value={environmentMock}>
-            <RouteContext.Provider value={mockedRouteContext}>
+            <RouteContext.Provider value={routerMock}>
               <AppearanceProvider appearanceKey={'signIn'}>
                 <FlowMetadataProvider flow={componentName as any}>
                   <InternalThemeProvider>
